syllogism_checker/logic_checker.py
import re
from typing import List

def check_syllogism(syllogism: str) -> List[str]:
    """
    Check the given syllogism for formal structure errors.
    Returns a list of errors found.
    """
    errors = []
    
    # Split the syllogism into premises and conclusion
    parts = syllogism.split('\n')
    if len(parts) != 3:
        errors.append("A syllogism must have exactly three parts: two premises and one conclusion.")
        return errors
    
    premise1, premise2, conclusion = parts
    
    # Check for basic structure errors
    if not premise1.endswith('.'):
        errors.append("The first premise must end with a period.")
    if not premise2.endswith('.'):
        errors.append("The second premise must end with a period.")
    if not conclusion.endswith('.'):
        errors.append("The conclusion must end with a period.")
    
    # Check for logical consistency (simplified example)
    terms = re.findall(r'\b\w+\b', syllogism)
    unique_terms = set(terms)
    if len(unique_terms) < 3:
        errors.append("A valid syllogism must contain at least three unique terms.")
    
    return errors